refactor: model -> domain 패키지 명 수정
